// src/app/classes/app-state.interface.ts
export interface ResultEntry {
  [prop: string]: any;
  response: APIResponse;
}

export interface ModelState<T = any> {
  byId: LookupTable<T>;
  loading: LookupTable<boolean>;
  results: LookupTable<ResultEntry>;
}

export interface GroupState extends ModelState<Group> {
  members: LookupTable<string[]>;
}
// ...
export interface LookupTable<T> {
  [id: string]: T;
}